+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [20.0.0] 2019-10-24
+### Added
+- Added "exists" support to skip logic
+### Changes
+- The trigger.code and triger.value.code were "hidden" features, but the issue of code system 
+  wasn't taken into consideration and therefore was broken.
+  This has been reimplemented so that when the source question is of type CNE or CWE,  
+  the trigger.value will be a hash with any or all of these three fields:
+  code, system, and text, and the trigger matching will be based on these fields. Please
+  see the updated lforms definition for more details. 
+
 ## [19.0.2] 2019-11-04
 ### Fixed
 - Fixed a bug that autocompleter options are not set for items in templateOptions.